1366: Refactor shared import form code into a trait
